refactor(today): CartPanelとuseMealLogを切り出す
TodayClient のカート表示と食事ログ（日付・一覧・削除・再取得）を
設計書 Issue #218 §3.2–3.3 に沿って分離する。挙動は変えない。

closes #220
related to #213

Made-with: Cursor
